Split dataManager
